#261: Add CessdaMetadataValidatorFactory.splitListRecordsResponse(), a function to split a single ListRecords response into multiple documents to be validated
This method takes an XML InputSource, detects if the document given is a ListRecords response, and splits the document into individual Document objects that can then be validated. The URI of the split documents is set to `InputSource.getSystemId() + '#' + identifier`. The method throws IllegalArgumentException if the XML document is not a ListRecords response.